Feature/trim endpoints (#155)
* Move things around on the api

* Remove endOfNight endpoint
Remove drivesr/me endpoint
Remove post /drivers/{id}/rides for assigning requests.
Update endpoint for a driver to get ride assigned to them

* Progress on ride request controller

* Fix ride request issues

* Remove driver location endpoint. Include driver location when fetching driver

* Return latest ride request for driver as part of the driver object. Tagged possibly deprecated api calls with console.log

* Remove driver status. Can be infered from the latest ride request's status

* Replace the need for the driver's rides lookup by using just the latest ride request assigned to the driver

* Remove ride cancellation endpoint. Allow rider to simply update the ride request record (just the status).

* Allow driver to cancel ride if the driver requested a ride

* Remove unused user registration service. Changes SafeRides to Safe Rides

* Fix multiple cas redirects

* http -> https

* Cleanup

+++ safe-rides-api/src/main/java/edu/csus/asi/saferides/model/RideRequest.java
@@ -1,12 +1,8 @@
 package edu.csus.asi.saferides.model;
 
-import edu.csus.asi.saferides.security.model.User;
 import edu.csus.asi.saferides.utility.Util;
-import org.hibernate.annotations.*;
-import org.hibernate.annotations.CascadeType;
 
 import javax.persistence.*;
-import javax.persistence.Entity;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 
@@ -80,13 +76,13 @@ public class RideRequest {
      * The odometer reading of the driver's vehicle at the beginning of the ride
      */
     @Column
-    private int startOdometer;
+    private long startOdometer;
 
     /**
      * The odometer reading of the driver's vehicle at the time of ride completion
      */
     @Column
-    private int endOdometer;
+    private long endOdometer;
 
     /**
      * Line 1 of rider's pickup address
@@ -379,7 +375,7 @@ public class RideRequest {
      *
      * @return driver's start odometer at the time of assignment
      */
-    public int getStartOdometer() {
+    public long getStartOdometer() {
         return startOdometer;
     }
 
@@ -388,7 +384,7 @@ public class RideRequest {
      *
      * @param startOdometer the driver's start odometer at the time of assignment
      */
-    public void setStartOdometer(int startOdometer) {
+    public void setStartOdometer(long startOdometer) {
         this.startOdometer = startOdometer;
     }
 
@@ -397,7 +393,7 @@ public class RideRequest {
      *
      * @return the driver's end odometer at the time of ride completion
      */
-    public int getEndOdometer() {
+    public long getEndOdometer() {
         return endOdometer;
     }
 
@@ -406,7 +402,7 @@ public class RideRequest {
      *
      * @param endOdometer the driver's end odometer at the time of ride completion
      */
-    public void setEndOdometer(int endOdometer) {
+    public void setEndOdometer(long endOdometer) {
         this.endOdometer = endOdometer;
     }
